[Refactor] - Extract websocket options to a new class (#6461)
